URL정리, 파일분리
