+++ html/create.html
@@ -1,4 +1,4 @@
-<form action="/create_table" method="post">
+<form action="/table/create_table" method="post">
   <p><input type="text" name="title" id="user_title" placeholder="구분" class="bar"></p>
   <p><textarea name="available" id="user_able" placeholder="가능여부" class="bar"></textarea></p>
   <p><input type="submit" value="Upload" class="bar"><br></p>
