+++ html/join.html
@@ -1,4 +1,4 @@
-<form action="/join_event" method="post">
+<form action="/event/join_event" method="post">
   <p><input type="text" name="id" id="user_title" placeholder="id" class="bar"></p>
   <p><input type="submit" value="Join" class="bar"><br></p>
 </form>
