+++ html/main.html
@@ -1,3 +1,3 @@
-<p><script>document.write(u_id);</script> | <a href="/logout">logout</a></p>
-<p><a href="/create">create new timetable</a></p>
-<p><a href="/make">make new event</a> | <a href="/join">join new event</a> </p>
+<p><script>document.write(u_id);</script> | <a href="/auth/logout">logout</a></p>
+<p><a href="/table/create">create new timetable</a></p>
+<p><a href="/event/make">make new event</a> | <a href="/event/join">join new event</a> </p>
