+++ html/make.html
@@ -1,4 +1,4 @@
-<form action="/make_event" method="post">
+<form action="/event/make_event" method="post">
   <p><input type="text" name="title" id="user_title" placeholder="제목" class="bar"></p>
   <p><input type="submit" value="Make" class="bar"><br></p>
 </form>
