+++ html/register.html
@@ -1,5 +1,5 @@
 <strong>절대 평소 사용하는 비밀번호를 사용하지 마세요. 보안에 취약한 사이트입니다.</strong>
-<form action="/register_process" method="post">
+<form action="/auth/register_process" method="post">
   <p><input type="text" name="id" placeholder="아이디" required></p>
   <p><input type="password" name="pw" placeholder="비밀번호" required></p>
   <p><input type="password" name="pw2" placeholder="비밀번호 확인" required></p>
