+++ js/db.js
@@ -1,0 +1,6 @@
+var low = require('lowdb');
+var FileSync = require('lowdb/adapters/FileSync');
+var adapter = new FileSync('db.json');
+var db = low(adapter);
+db.defaults({users:[], topics:[]}).write();
+module.exports = db;
