+++ js/lib.js
@@ -1,0 +1,5 @@
+module.exports = {
+  authIsOwner:function(request, response){
+    return request.user
+  }
+}
